requirements.txt
fastapi
google-generativeai
PyPDF2
python-multipart
uvicorn
requests
python-dotenv
